Refactor to eliminate code duplication.
